requirements.txt
Streamlit==1.37.1
Ultralytics==8.3.27
OpenCV-python==4.10.0
NumPy==1.26.4
Pillow==10.4.0